+++ HISTORY.md
@@ -26,6 +26,7 @@
 
 * Deprecates `Group.create_virtual` in favor of `VirtualGroup.create`.
 * Deprecates `NetCDF4ConverterEngine.add_scalar_dim_converter` in favor of `NetCDF4ConverterEngine.add_scalar_to_dim_converter`.
+* Deprecates `add_array` in favor of `add_array_creator`.
 
 ## TileDB-CF-Py Release 0.3.0